Update evaluate.py
In the method initialization, you are expecting model_pair but while using this method you are sending only strong model which is throwing a runtime error while running the code. After making this change in my local it was working fine. Please check this from your end as well

Thank you
Keerthy Kaushik